made system to use verilog

+++ last_translator.py
@@ -61,7 +61,7 @@ class LastTranslator(ast.NodeTransformer):
         return_var: VariableDict,
         return_type: VariableDict,
     ) -> None:
-        self.precision = 16
+        self.precision = 48
         self.array_list = arrat_list
         self.float_list = float_list
         self.complex_list = complex_list
